pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nearby-fact-bot"
version = "1.0.0"
description = "Telegram bot that provides interesting facts about places near your location"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[webhooks]==21.7",
    "openai>=1.54.0",
    "asyncio-throttle>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "black>=24.3.0",
    "pytest>=8.0.0",
    "pytest-anyio>=0.0.0",
    "pytest-mock>=3.12.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101"]  # assert allowed in tests

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 